Nutrition macros int to double

+++ src/Services/NutritionService/GymApp.NutritionService.Data/Entities/Diet.cs
@@ -6,10 +6,10 @@ public class Diet
 {
     public int Id { get; set; }
     public string? Name { get; set; }
-    public int Calories => DietMeals.Sum(dm => dm.Meal.Calories);
-    public int Protein => DietMeals.Sum(dm => dm.Meal.Protein);
-    public int Carbohydrates => DietMeals.Sum(dm => dm.Meal.Carbohydrates);
-    public int Fats => DietMeals.Sum(dm => dm.Meal.Fats);
+    public double Calories => DietMeals.Sum(dm => dm.Meal.Calories);
+    public double Protein => DietMeals.Sum(dm => dm.Meal.Protein);
+    public double Carbohydrates => DietMeals.Sum(dm => dm.Meal.Carbohydrates);
+    public double Fats => DietMeals.Sum(dm => dm.Meal.Fats);
 
     public ICollection<DietMeal> DietMeals { get; set; } = null!;
 }
