+++ src/Services/NutritionService/GymApp.NutritionService.Data/Entities/Meal.cs
@@ -6,10 +6,10 @@ public class Meal
 {
     public int Id { get; set; }
     public string? Name { get; set; }
-    public int Calories => MealFoods.Sum(mf => mf.Food.Calories);
-    public int Protein => MealFoods.Sum(mf => mf.Food.Protein);
-    public int Carbohydrates => MealFoods.Sum(mf => mf.Food.Carbohydrates);
-    public int Fats => MealFoods.Sum(mf => mf.Food.Fats);
+    public double Calories => MealFoods.Sum(mf => mf.Food.Calories);
+    public double Protein => MealFoods.Sum(mf => mf.Food.Protein);
+    public double Carbohydrates => MealFoods.Sum(mf => mf.Food.Carbohydrates);
+    public double Fats => MealFoods.Sum(mf => mf.Food.Fats);
 
     public ICollection<MealFood> MealFoods { get; set; } = null!;
     public ICollection<DietMeal> DietMeals { get; set; } = null!;
